Add new store for tracking managed files.
This fixes an issue with removing the dangling links when the target is
a very large nested structure. The previous implementation would end up
scanning every single file which can take a very long time...

The new implementation only scans files that we have created links for.
This is potentially going to end up missing stuff and won't work for
existing links that have been generated. So far unlinking only updates
this store and doesn't actually use the store to find files to unlink.